+++ packages/spacecat-shared-data-access/test/it/db.test.js
@@ -82,14 +82,17 @@ const TEST_DA_CONFIG = {
   tableNameOrganizations: 'spacecat-services-organizations',
   tableNameSites: 'spacecat-services-sites',
   tableNameSiteCandidates: 'spacecat-services-site-candidates',
+  tableNameConfigurations: 'spacecat-services-configurations',
   indexNameAllSites: 'spacecat-services-all-sites',
   indexNameAllSitesOrganizations: 'spacecat-services-all-sites-organizations',
   indexNameAllOrganizations: 'spacecat-services-all-organizations',
+  indexNameAllOrganizationsByImsOrgId: 'spacecat-services-all-organizations-by-ims-org-id',
   indexNameAllSitesByDeliveryType: 'spacecat-services-all-sites-by-delivery-type',
   indexNameAllLatestAuditScores: 'spacecat-services-all-latest-audit-scores',
   pkAllSites: 'ALL_SITES',
   pkAllOrganizations: 'ALL_ORGANIZATIONS',
   pkAllLatestAudits: 'ALL_LATEST_AUDITS',
+  pkAllConfigurations: 'ALL_CONFIGURATIONS',
 };
 
 describe('DynamoDB Integration Test', async () => {
@@ -132,6 +135,22 @@ describe('DynamoDB Integration Test', async () => {
     dynamoDbLocalProcess.kill();
   });
 
+  it('gets configuration by Version', async () => {
+    const configuration = await dataAccess.getConfigurationByVersion('v1');
+
+    expect(configuration).to.be.an('object');
+
+    expect(configuration.getVersion()).to.equal('v1');
+  });
+
+  it('gets configuration', async () => {
+    const configuration = await dataAccess.getConfiguration();
+
+    expect(configuration).to.be.an('object');
+
+    expect(configuration.getVersion()).to.equal('v2');
+  });
+
   it('gets organizations', async () => {
     const organizations = await dataAccess.getOrganizations();
 
@@ -150,6 +169,16 @@ describe('DynamoDB Integration Test', async () => {
 
     checkOrganization(organization);
     expect(organization.getId()).to.equal(orgId);
+  });
+
+  it('gets organization by IMS Org ID', async () => {
+    const imsOrgId = (await dataAccess.getOrganizations())[0].getImsOrgId();
+    const organization = await dataAccess.getOrganizationByImsOrgID(imsOrgId);
+
+    expect(organization).to.be.an('object');
+
+    checkOrganization(organization);
+    expect(organization.getImsOrgId()).to.equal(imsOrgId);
   });
 
   it('adds a new organization', async () => {